public memory ready, trace wip
* refactor into cairo_interop.rs
* document and cleanup code
* pad trace to power of 2
